requirements.txt
asgiref==3.5.1
colorama==0.4.4
Django==4.0.4
Pillow==9.1.0
qrcode==7.3.1
sqlparse==0.4.2
tzdata==2022.1
